more response cleanup
