+++ core/src/test/java/org/apache/iceberg/rest/responses/TestFetchPlanningResultResponseParser.java
@@ -30,7 +30,6 @@ import com.fasterxml.jackson.databind.JsonNode;
 import java.util.List;
 import org.apache.iceberg.BaseFileScanTask;
 import org.apache.iceberg.DeleteFile;
-import org.apache.iceberg.FetchPlanningResultResponseParser;
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.PartitionSpecParser;
 import org.apache.iceberg.SchemaParser;
@@ -39,7 +38,7 @@ import org.apache.iceberg.expressions.ResidualEvaluator;
 import org.apache.iceberg.rest.PlanStatus;
 import org.junit.jupiter.api.Test;
 
-public class TestFetchPlanningResultResponse {
+public class TestFetchPlanningResultResponseParser {
 
   @Test
   public void nullAndEmptyCheck() {
@@ -57,7 +56,7 @@ public class TestFetchPlanningResultResponse {
     assertThatThrownBy(
             () ->
                 FetchPlanningResultResponseParser.toJson(
-                    new FetchPlanningResultResponse.Builder().build()))
+                    FetchPlanningResultResponse.builder().build()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage("Invalid status: null");
 
@@ -79,7 +78,7 @@ public class TestFetchPlanningResultResponse {
   public void roundTripSerdeWithValidSubmittedStatus() {
     PlanStatus planStatus = PlanStatus.fromName("submitted");
     FetchPlanningResultResponse response =
-        new FetchPlanningResultResponse.Builder().withPlanStatus(planStatus).build();
+        FetchPlanningResultResponse.builder().withPlanStatus(planStatus).build();
 
     String expectedJson = "{\"plan-status\":\"submitted\"}";
     String json = FetchPlanningResultResponseParser.toJson(response);
@@ -92,13 +91,12 @@ public class TestFetchPlanningResultResponse {
   @Test
   public void roundTripSerdeWithInvalidPlanStatusSubmittedWithTasksPresent() {
     PlanStatus planStatus = PlanStatus.fromName("submitted");
-    FetchPlanningResultResponse response =
-        new FetchPlanningResultResponse.Builder()
-            .withPlanStatus(planStatus)
-            .withPlanTasks(List.of("task1", "task2"))
-            .build();
-
-    assertThatThrownBy(() -> FetchPlanningResultResponseParser.toJson(response))
+    assertThatThrownBy(
+            () ->
+                FetchPlanningResultResponse.builder()
+                    .withPlanStatus(planStatus)
+                    .withPlanTasks(List.of("task1", "task2"))
+                    .build())
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage("Invalid response: tasks can only be returned in a 'completed' status");
 
@@ -113,13 +111,12 @@ public class TestFetchPlanningResultResponse {
   @Test
   public void roundTripSerdeWithInvalidPlanStatusSubmittedWithDeleteFilesNoFileScanTasksPresent() {
     PlanStatus planStatus = PlanStatus.fromName("submitted");
-    FetchPlanningResultResponse response =
-        new FetchPlanningResultResponse.Builder()
-            .withPlanStatus(planStatus)
-            .withDeleteFiles(List.of(FILE_A_DELETES))
-            .build();
-
-    assertThatThrownBy(() -> FetchPlanningResultResponseParser.toJson(response))
+    assertThatThrownBy(
+            () ->
+                FetchPlanningResultResponse.builder()
+                    .withPlanStatus(planStatus)
+                    .withDeleteFiles(List.of(FILE_A_DELETES))
+                    .build())
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage(
             "Invalid response: deleteFiles should only be returned with fileScanTasks that reference them");
@@ -151,7 +148,7 @@ public class TestFetchPlanningResultResponse {
 
     PlanStatus planStatus = PlanStatus.fromName("completed");
     FetchPlanningResultResponse response =
-        new FetchPlanningResultResponse.Builder()
+        FetchPlanningResultResponse.builder()
             .withPlanStatus(planStatus)
             .withFileScanTasks(List.of(fileScanTask))
             .withDeleteFiles(List.of(FILE_A_DELETES))
@@ -176,7 +173,7 @@ public class TestFetchPlanningResultResponse {
     assertThat(json).isEqualTo(expectedToJson);
 
     // make an unbound json where you expect to not have partitions for the data file,
-    // delete files as service does not send parition spec
+    // delete files as service does not send partition spec
     String expectedFromJson =
         "{\"plan-status\":\"completed\","
             + "\"delete-files\":[{\"spec-id\":0,\"content\":\"POSITION_DELETES\","
@@ -193,7 +190,7 @@ public class TestFetchPlanningResultResponse {
     FetchPlanningResultResponse fromResponse = FetchPlanningResultResponseParser.fromJson(json);
     // Need to make a new response with partitionSpec set
     FetchPlanningResultResponse copyResponse =
-        new FetchPlanningResultResponse.Builder()
+        FetchPlanningResultResponse.builder()
             .withPlanStatus(fromResponse.planStatus())
             .withPlanTasks(fromResponse.planTasks())
             .withDeleteFiles(fromResponse.deleteFiles())
